+++ mmex.net.core/Entities/Payee.cs
@@ -2,9 +2,9 @@ namespace mmex.net.core.Entities;
 
 public class Payee
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public int? CategoryId { get; set; }
+    public long? CategoryId { get; set; }
     public string? Number { get; set; }
     public string? Website { get; set; }
     public string? Notes { get; set; }
