+++ mmex.net.core/Entities/ScheduledTransaction.cs
@@ -5,17 +5,17 @@ namespace mmex.net.core.Entities;
 /// <summary>Maps BILLSDEPOSITS_V1. Same structure as Transaction plus scheduling fields.</summary>
 public class ScheduledTransaction
 {
-    public int Id { get; set; }
-    public int AccountId { get; set; }
-    public int? ToAccountId { get; set; }
-    /// <summary>C++ stores -1 or 0 for "no payee". Treat &lt;= 0 as null.</summary>
-    public int PayeeId { get; set; }
+    public long Id { get; set; }
+    public long AccountId { get; set; }
+    public long? ToAccountId { get; set; }
+    /// <summary>C++ stores -1 for "no payee". Treat &lt;= 0 as null.</summary>
+    public long PayeeId { get; set; }
     public TransactionType Type { get; set; }
     public decimal Amount { get; set; }
     public TransactionStatus Status { get; set; }
     public string? Number { get; set; }
     public string? Notes { get; set; }
-    public int? CategoryId { get; set; }
+    public long? CategoryId { get; set; }
     /// <summary>Stored as ISO 8601 TEXT "YYYY-MM-DD".</summary>
     public string? Date { get; set; }
     public int? FollowUpId { get; set; }
