+++ mmex.net.core/Entities/ShareInfo.cs
@@ -3,8 +3,8 @@ namespace mmex.net.core.Entities;
 /// <summary>Maps SHAREINFO_V1 — share details for investment transactions.</summary>
 public class ShareInfo
 {
-    public int Id { get; set; }
-    public int CheckingAccountId { get; set; }
+    public long Id { get; set; }
+    public long CheckingAccountId { get; set; }
     public decimal? ShareNumber { get; set; }
     public decimal? SharePrice { get; set; }
     public decimal? ShareCommission { get; set; }
