+++ mmex.net.core/Entities/TagLink.cs
@@ -2,11 +2,11 @@ namespace mmex.net.core.Entities;
 
 public class TagLink
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     /// <summary>e.g. "Transaction", "Stock", "Asset".</summary>
     public string RefType { get; set; } = string.Empty;
-    public int RefId { get; set; }
-    public int TagId { get; set; }
+    public long RefId { get; set; }
+    public long TagId { get; set; }
 
     // Navigation
     public Tag? Tag { get; set; }
