+++ mmex.net.core/Services/IAccountService.cs
@@ -7,10 +7,10 @@ public interface IAccountService
     Task<IList<Account>> GetAllAsync();
     Task<IList<Account>> GetAllOpenAsync();
     Task<IList<Account>> GetFavoritesAsync();
-    Task<Account?> GetByIdAsync(int id);
+    Task<Account?> GetByIdAsync(long id);
     Task<Account> CreateAsync(Account account);
     Task<Account> UpdateAsync(Account account);
-    Task DeleteAsync(int id);
-    Task<decimal> GetBalanceAsync(int accountId);
-    Task<decimal> GetBalanceAtDateAsync(int accountId, DateOnly date);
+    Task DeleteAsync(long id);
+    Task<decimal> GetBalanceAsync(long accountId);
+    Task<decimal> GetBalanceAtDateAsync(long accountId, DateOnly date);
 }
