+++ mmex.net.core/Services/IPayeeService.cs
@@ -5,8 +5,8 @@ namespace mmex.net.core.Services;
 public interface IPayeeService
 {
     Task<IList<Payee>> GetAllAsync(bool activeOnly = true);
-    Task<Payee?> GetByIdAsync(int id);
+    Task<Payee?> GetByIdAsync(long id);
     Task<Payee> CreateAsync(Payee payee);
     Task<Payee> UpdateAsync(Payee payee);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
