+++ mmex.net.core/Services/IScheduledTransactionService.cs
@@ -6,13 +6,9 @@ public interface IScheduledTransactionService
 {
     Task<IList<ScheduledTransaction>> GetAllAsync();
     Task<IList<ScheduledTransaction>> GetDueAsync(DateOnly asOfDate);
-    Task<ScheduledTransaction?> GetByIdAsync(int id);
+    Task<ScheduledTransaction?> GetByIdAsync(long id);
     Task<ScheduledTransaction> CreateAsync(ScheduledTransaction scheduled);
     Task<ScheduledTransaction> UpdateAsync(ScheduledTransaction scheduled);
-    Task DeleteAsync(int id);
-    /// <summary>
-    /// Posts the scheduled transaction to CHECKINGACCOUNT_V1 and advances the schedule.
-    /// Returns the newly created transaction.
-    /// </summary>
-    Task<Transaction> ExecuteAsync(int scheduledId);
+    Task DeleteAsync(long id);
+    Task<Transaction> ExecuteAsync(long scheduledId);
 }
